requirements.txt
aws-cdk-lib==2.111.0
constructs>=10.0.0,<11.0.0
Flask==3.0.0
aws-wsgi==0.2.7
pynamodb==5.5.1
openai==1.6.0
weaviate-client==4.4b1